backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-socketio==5.8.0
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
eventlet==0.33.3
